Remove Users and split into two receive loops
Users contained a lot of mutexes which made it easy to create deadlocks. By
rather keeping the user object in the local stack we prevent these errors. The
chat messages will rather be handled by a separate goroutine using channels in
the future.

The two loops are introduced to separate concerns:

1. The first loop will handle all unauthenticated messages.
2. The second loop will handle all authenticated messages when the user object is set.