Add safe websock send/recv API
Previously one could send any type because `websock.Msg.Send` takes a
`interface {}`. However within the `marshalMessage` function the first
assertion makes sure that we have a `*Message`. This commit enables the static
type checker to warn us if we try to pass a object which is not a `*Message` to
send/recv.